refactor profile component
